requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
supabase>=2.4.0
openai>=1.30.0
anthropic>=0.28.0
google-genai>=0.5.0
google-auth>=2.29.0
google-api-python-client>=2.130.0
pillow>=10.3.0
pypdf>=4.2.0
python-docx>=1.1.0
auto-editor>=25.0.0
